#49 SetlistEditでInputValueを使えるようにする
